+++ packages/expo-router/src/exports.ts
@@ -47,3 +47,14 @@ export type { RedirectConfig } from './getRoutesCore';
 export type { SingularOptions } from './useScreens';
 
 export type * from './types';
+
+export {
+  Badge,
+  BadgeProps,
+  Icon,
+  IconProps,
+  Label,
+  LabelProps,
+  VectorIcon,
+  VectorIconProps,
+} from './primitives';