+++ packages/expo-router/src/native-tabs/NativeTabTrigger.tsx
@@ -2,27 +2,27 @@
 
 import { useNavigation, useRoute } from '@react-navigation/native';
 import { isValidElement, type ReactElement, type ReactNode } from 'react';
-import type { ImageSourcePropType } from 'react-native';
+import { type ImageSourcePropType } from 'react-native';
 
 import { NativeTabsTriggerTabBar } from './NativeTabsTriggerTabBar';
-import type {
-  NativeTabOptions,
-  NativeTabsTriggerTabBarProps,
-  NativeTabTriggerProps,
+import {
+  type NativeTabOptions,
+  type NativeTabsTriggerTabBarProps,
+  type NativeTabTriggerProps,
 } from './types';
 import { filterAllowedChildrenElements, isChildOfType } from './utils';
 import { useIsPreview } from '../link/preview/PreviewRouteContext';
+import type { VectorIconProps } from '../primitives';
 import { useSafeLayoutEffect } from '../views/useSafeLayoutEffect';
 import {
-  Icon,
-  Badge,
-  Label,
-  type LabelProps,
-  type IconProps,
-  type BadgeProps,
+  NativeTabsTriggerIcon,
+  NativeTabsTriggerBadge,
+  NativeTabsTriggerLabel,
+  NativeTabsTriggerVectorIcon,
+  type NativeTabsTriggerBadgeProps,
+  type NativeTabsTriggerLabelProps,
+  type NativeTabsTriggerIconProps,
   type SourceIconCombination,
-  VectorIcon,
-  type VectorIconProps,
 } from './common/elements';
 
 /**
@@ -55,7 +55,7 @@ import {
  *   return (
  *     <View>
  *       <NativeTabs.Trigger>
- *         <Label>Home</Label>
+ *         <NativeTabs.Trigger.Label>Home</NativeTabs.Trigger.Label>
  *       </NativeTabs.Trigger>
  *       <Text>This is home screen!</Text>
  *     </View>
@@ -91,6 +91,10 @@ function NativeTabTriggerImpl(props: NativeTabTriggerProps) {
 
 export const NativeTabTrigger = Object.assign(NativeTabTriggerImpl, {
   TabBar: NativeTabsTriggerTabBar,
+  Label: NativeTabsTriggerLabel,
+  Icon: NativeTabsTriggerIcon,
+  Badge: NativeTabsTriggerBadge,
+  VectorIcon: NativeTabsTriggerVectorIcon,
 });
 
 export function convertTabPropsToOptions(
@@ -120,18 +124,18 @@ export function convertTabPropsToOptions(
         nativeProps: unstable_nativeProps,
       };
   const allowedChildren = filterAllowedChildrenElements(children, [
-    Badge,
-    Label,
-    Icon,
+    NativeTabsTriggerBadge,
+    NativeTabsTriggerLabel,
+    NativeTabsTriggerIcon,
     NativeTabsTriggerTabBar,
   ]);
   return allowedChildren.reduce<NativeTabOptions>(
     (acc, child) => {
-      if (isChildOfType(child, Badge)) {
+      if (isChildOfType(child, NativeTabsTriggerBadge)) {
         appendBadgeOptions(acc, child.props);
-      } else if (isChildOfType(child, Label)) {
+      } else if (isChildOfType(child, NativeTabsTriggerLabel)) {
         appendLabelOptions(acc, child.props);
-      } else if (isChildOfType(child, Icon)) {
+      } else if (isChildOfType(child, NativeTabsTriggerIcon)) {
         appendIconOptions(acc, child.props);
       } else if (isChildOfType(child, NativeTabsTriggerTabBar)) {
         appendTabBarOptions(acc, child.props);
@@ -142,7 +146,7 @@ export function convertTabPropsToOptions(
   );
 }
 
-function appendBadgeOptions(options: NativeTabOptions, props: BadgeProps) {
+function appendBadgeOptions(options: NativeTabOptions, props: NativeTabsTriggerBadgeProps) {
   if (props.children) {
     options.badgeValue = String(props.children);
     options.selectedBadgeBackgroundColor = props.selectedBackgroundColor;
@@ -154,7 +158,7 @@ function appendBadgeOptions(options: NativeTabOptions, props: BadgeProps) {
   }
 }
 
-function appendLabelOptions(options: NativeTabOptions, props: LabelProps) {
+function appendLabelOptions(options: NativeTabOptions, props: NativeTabsTriggerLabelProps) {
   if (props.hidden) {
     options.title = '';
   } else {
@@ -165,7 +169,7 @@ function appendLabelOptions(options: NativeTabOptions, props: LabelProps) {
   }
 }
 
-export function appendIconOptions(options: NativeTabOptions, props: IconProps) {
+export function appendIconOptions(options: NativeTabOptions, props: NativeTabsTriggerIconProps) {
   if ('src' in props && props.src) {
     const icon = convertIconSrcToIconOption(props);
     options.icon = icon?.icon;
@@ -224,7 +228,7 @@ function convertIconSrcToIconOption(
 function convertSrcOrComponentToSrc(src: ImageSourcePropType | ReactElement | undefined) {
   if (src) {
     if (isValidElement(src)) {
-      if (src.type === VectorIcon) {
+      if (src.type === NativeTabsTriggerVectorIcon) {
         const props = src.props as VectorIconProps<string>;
         return { src: props.family.getImageSource(props.name, 24, 'white') };
       } else {